Be more silent when doing one off command
Don't spam the same things all the time when doing poor mans automation
in shell scripts. If the users really wants to see status, then just
insert `m2ee status` as extra command.

This also already helps the case where a users wants to do `m2ee stop`
when a runtime process is so borked that calls to the admin API hang.