nagios: use timeout for all admin api calls
Instead of using no timeout (the default), set sensible low timeouts for
collecting information used in the checks.

The total timeout of running either of them is not the same as the sum
of all individual ones. The timeouts are there to make sure the whole
process aborts if any of them are hit.

Also, also add M2EEAdminHTTPException and handle it as if the admin api
is unavailable. This error should never happen.