No longer ask to create a database
The whole DTAP setting is deprecated, let's get rid of its usage.

The program will no longer try to use builtin functionality in the
Mendix runtime to create a database.

Running the Mendix Runtime with a database user with admin privileges to
create databases (and probably also remove them) is a very bad idea
anyway.