Stop using DTAPmode and Constants from config.json
As stated in the documentation of the full example config file, doing
this and using the DTAPMode setting with a value 'D' or 'T' has been
deprecated for a long time already.

Remove this functionality now to lower code complexity. The config.json
style config should only be used when directly running from the Modeler.

Note that the _merge_microflow_constants also merged other runtime
configuration settings from the config.json file. Yuck.